+++ backend-lib/src/main/java/cash/z/ecc/android/sdk/internal/jni/FakeRustBackend.kt
@@ -144,6 +144,15 @@ class FakeRustBackend(
         error("Intentionally not implemented yet.")
     }
 
+    override suspend fun addSaplingProofGenerationKeys(
+        pczt: ByteArray,
+        ufvk: String,
+        externalNsk: ByteArray,
+        internalNsk: ByteArray
+    ): ByteArray {
+        error("Intentionally not implemented yet.")
+    }
+
     override suspend fun pcztRequiresSaplingProofs(pczt: ByteArray): Boolean {
         error("Intentionally not implemented yet.")
     }